pyproject.toml
[tool.poetry]
name = "kursovaya"
version = "0.1.0"
description = ""
authors = ["ruslankh <ruselkotya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
load-dotenv = "^0.1.0"
requests = "^2.32.3"
pytest = "^8.3.5"
freezegun = "^1.5.1"
openpyxl = "^3.1.5"
datetime = "^5.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''
[tool.isort]
# максимальная длина строки
line_length = 119